backend/requirements.txt
fastapi
uvicorn
motor
python-dotenv
google-genai
pydantic
pydantic-settings
dnspython
groq
